redactra/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.11.7
pydantic-settings==2.7.0
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib==1.7.4
cryptography==44.0.0
email-validator==2.2.0
pypdf==5.1.0
Pillow==11.0.0
pytesseract==0.3.13
reportlab==4.2.5
openpyxl==3.1.5
# Optional: only if using Supabase storage (omit for local file storage)
# supabase==2.18.1
# Optional: only if using PostgreSQL (omit for SQLite-only)
# psycopg2-binary==2.9.10
apscheduler==3.10.4
